Complete assignment
Sorts array of args and uses the .every method of arrays to check each
array for negative or non number vals and set a boolean for each for
error handling. Using a for lops, iterates from the last val of theArgs
to the first, adding them to the sum var, then returning the final sum.